Spread code ouut nicer, add debugging print
